DB migration script

+++ internal/db/migrations/000005_pet_recommend_schema.up.sql
@@ -1,7 +1,7 @@
 create table pet_recommend(
 		id1 varbinary(16) NOT NULL PRIMARY KEY,
 		id2 varbinary(16) NOT NULL,
-		score float(55),
-		status int
+		score float(20),
+		status int,
 		FOREIGN KEY(id2) REFERENCES pet(id)
 	);
